Add support for aborting a zip operation.
`ProgressCallback.h` was changed to use pure virtual functions to eliminate a compile error on the UE4 library side.

`ArchiveExtractCallback` / `ArchiveUpdateCallback`
Implementation of `CheckBreak()`. The primary means of allowing zip operations to break is to return `CheckBreak()` in the vast majority of interface callbacks that an archive gets over its lifetime. As to which callbacks should return `CheckBreak()` vs not, this was mostly copied from the SevenZip sample code. I did omit the check from `GetProperty` because it seemed a little overkill and I was also mildly (perhaps without reason) worried about hitting the mutex on the `FThreadSafeBool` too frequently since `GetProperty` seemed high on the most frequently called list.

Since this is COM, returning `E_ABORT` in any of these callbacks effectively terminates the operation. `CheckBreak()` is simply an easy entry point that in-turn calls through to the `ProgressCallback` virtual method `OnCheckBreak()` which is overriden in the UE4 library implementation using an `FThreadSafeBool`.