Fix file-rename operations and Allow format detection to fail
Finally got the 7zpp-TestApp running and noticed that extraction test 4 was failing. My fault, I was stripping the path seperation character as part of the invalid character check. Fixed that.

I noticed that `DetectCompressionFormat` always returns true, even if the format is undetected at the end, yet there are many places in the UE4 plugin that assume a false state. It now returns false if no format could be detected.